+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - [TSLint] Bump tslint from 6.0.0 to 6.1.2 [#1008](https://github.com/sider/runners/pull/1008)
 - [GolangCI-Lint] Bump golangci/golangci-lint from v1.24.0 to v1.25.1 [#1023](https://github.com/sider/runners/pull/1023)
 - [FxCop] Bump FxCop from 2.9.8 to 3.0.0 [#1020](https://github.com/sider/runners/pull/1020)
+- [hadolint] Bump hadolint/hadolint from v1.17.5-debian to v1.17.6-debian [#1025](https://github.com/sider/runners/pull/1025)
 
 ## 0.22.4